third_party/blink/web_tests/svg/dynamic-updates/SVGPatternElement-svgdom-patternTransform-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'patternTransform' property of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS transform.type is SVGTransform.SVG_TRANSFORM_MATRIX
PASS transform.matrix.a is 1.0
PASS transform.matrix.b is 0.0
PASS transform.matrix.c is 0.0
PASS transform.matrix.d is 1.0
PASS transform.matrix.e is 0.0
PASS transform.matrix.f is 0.0
PASS transform.angle is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.a is 1.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.b is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.c is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.d is 1.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.e is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.f is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.a is 4.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.b is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.c is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.d is 1.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.e is 0.0
PASS patternElement.patternTransform.baseVal.getItem(0).matrix.f is 0.0
PASS successfullyParsed is true

TEST COMPLETE
